Introduce RawMessage
Replace RawBackendMessage and RawFrontendMessage with RawMessage and
change all places where they where used. It's now easier to just
read the message and extract it's type without converting it to its
"expanded" form. In passthru mode, it way easier to just pass the
message if the content is of no interest to us.

BackendMessages which where introduce not long ago aso disappear in
this commit.